Updated read me 

Fix delete

Deleted extra app.properties file

+++ src/main/java/com/id/recipes/recipes_api/service/RecipeServiceImpl.java
@@ -9,7 +9,6 @@ import com.id.recipes.recipes_api.model.rest.RecipeRequest;
 import com.id.recipes.recipes_api.repository.RecipeRepository;
 import com.id.recipes.recipes_api.utility.RecipeSpecification;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -56,12 +55,11 @@ public class RecipeServiceImpl implements RecipeService{
      */
     @Override
     public void deleteById(long recipeId) {
-        try {
-            recipeRepository.deleteById(recipeId);
-        }  catch (EmptyResultDataAccessException ex) {
-            throw new RecipeNotFoundException("Recipe not found for id=" + recipeId);
-        }
 
+        if (!recipeRepository.existsById(recipeId)) {
+            throw new RecipeNotFoundException("Recipe not found with id: " + recipeId);
+        }
+        recipeRepository.deleteById(recipeId);
     }
 
     /**
@@ -70,7 +68,6 @@ public class RecipeServiceImpl implements RecipeService{
      * @param recipeRequest the recipe request received
      * @return the recipe
      */
-    @SuppressWarnings("unchecked")
     @Override
     public Recipe createRecipe(RecipeRequest recipeRequest) {
         Recipe createRecipe = new Recipe();
